+++ trees-examples/src/main/java/com/mattunderscore/trees/examples/ReadmeExamples.java
@@ -51,8 +51,7 @@ public final class ReadmeExamples {
         final ServiceLoader<Trees> serviceLoader = ServiceLoader.load(Trees.class);
         final Trees trees = serviceLoader.iterator().next();
 
-        final SortingTreeBuilder<Integer> builder = trees.treeBuilders()
-            .sortingTreeBuilder(new ComparableComparator<Integer>());
+        final SortingTreeBuilder<Integer> builder = trees.treeBuilders().sortingTreeBuilder();
         final BinarySearchTree<Integer> tree = builder
             .addElement(2)
             .addElement(1)