+++ CHANGELOG.md
@@ -48,3 +48,7 @@ Thursday January 22 21:27:56 2015
 * Separated tree implementations into different modules
 * Separated tests that rely on multiple modules into separate module
 * Separated tree implementations into different packages
+* Made examples executable
+* Renamed examples module
+* Allow the creation of sorting and sorted tree builders without specifying a comparator if the element type is
+comparable